refactor(tag): clarify FETagType & TagType
One comes from Frontend, the other is our enhanced DCR one.
