+++ dotcom-rendering/src/amp/components/topMeta/SeriesLink.tsx
@@ -2,6 +2,7 @@ import { css } from '@emotion/react';
 import { headline } from '@guardian/source-foundations';
 import React from 'react';
 import { pillarPalette_DO_NOT_USE } from '../../../lib/pillars';
+import type { TagType } from '../../../types/tag';
 
 const seriesStyle = (pillar: ArticleTheme) => css`
 	color: ${pillarPalette_DO_NOT_USE[pillar].main};
