+++ dotcom-rendering/src/amp/lib/byline-tokens.ts
@@ -1,3 +1,5 @@
+import type { TagType } from '../../types/tag';
+
 export const bylineTokens = (
 	byline: string,
 	contributorTags: TagType[],
