+++ dotcom-rendering/src/amp/lib/tag-utils.ts
@@ -1,3 +1,5 @@
+import type { TagType } from '../../types/tag';
+
 export const filterForTagsOfType = (
 	tags: TagType[],
 	tagType: string,
