+++ dotcom-rendering/src/lib/age-warning.ts
@@ -1,3 +1,5 @@
+import type { TagType } from '../types/tag';
+
 export const getAgeWarning = (
 	tags: TagType[],
 	webPublicationDate: string,
