+++ dotcom-rendering/src/model/find-pillar.ts
@@ -1,4 +1,5 @@
 import { ArticlePillar, ArticleSpecial } from '@guardian/libs';
+import type { TagType } from '../types/tag';
 
 export const findPillar: (
 	name: string,
