+++ dotcom-rendering/src/web/components/ArticleTitle.tsx
@@ -2,6 +2,7 @@ import { css } from '@emotion/react';
 import { ArticleDesign, ArticleDisplay } from '@guardian/libs';
 import { from, until } from '@guardian/source-foundations';
 import type { BadgeType } from '../../types/badge';
+import type { TagType } from '../../types/tag';
 import { Badge } from './Badge';
 import { SeriesSectionLink } from './SeriesSectionLink';
