+++ dotcom-rendering/src/web/components/Contributor.tsx
@@ -4,6 +4,7 @@ import { from, headline, textSans, until } from '@guardian/source-foundations';
 import { getSoleContributor } from '../../lib/byline';
 import TwitterIcon from '../../static/icons/twitter.svg';
 import type { Palette } from '../../types/palette';
+import type { TagType } from '../../types/tag';
 import { interactiveLegacyClasses } from '../layouts/lib/interactiveLegacyStyling';
 import { decidePalette } from '../lib/decidePalette';
 import { BylineLink } from './BylineLink';
