+++ dotcom-rendering/src/web/components/LiveBlogEpic.importable.tsx
@@ -7,6 +7,7 @@ import type { EpicPayload } from '@guardian/support-dotcom-components/dist/dotco
 import React, { useEffect, useState } from 'react';
 import { createPortal } from 'react-dom';
 import { useArticleCounts } from '../../lib/articleCount';
+import type { TagType } from '../../types/tag';
 import { submitComponentEvent } from '../browser/ophan/ophan';
 import {
 	getLastOneOffContributionTimestamp,
