+++ dotcom-rendering/src/web/components/RichLinkComponent.importable.tsx
@@ -1,4 +1,5 @@
 import { ArticleDesign, ArticleDisplay, ArticleSpecial } from '@guardian/libs';
+import type { TagType } from '../../types/tag';
 import { decideFormat } from '../lib/decideFormat';
 import { useApi } from '../lib/useApi';
 import type { RichLinkImageData } from './RichLink';
