+++ dotcom-rendering/src/web/components/SeriesSectionLink.tsx
@@ -9,6 +9,7 @@ import {
 } from '@guardian/source-foundations';
 import type { BadgeType } from '../../types/badge';
 import type { Palette } from '../../types/palette';
+import type { TagType } from '../../types/tag';
 import { interactiveLegacyClasses } from '../layouts/lib/interactiveLegacyStyling';
 import { decidePalette } from '../lib/decidePalette';
 import { Badge } from './Badge';
