+++ dotcom-rendering/src/web/components/SignInGate/types.ts
@@ -1,3 +1,5 @@
+import type { TagType } from '../../../types/tag';
+
 export type CanShowGateProps = {
 	isSignedIn: boolean;
 	currentTest: CurrentSignInGateABTest;
