+++ dotcom-rendering/src/web/components/SlotBodyEnd/ReaderRevenueEpic.tsx
@@ -10,6 +10,7 @@ import type {
 	WeeklyArticleHistory,
 } from '@guardian/support-dotcom-components/dist/dotcom/src/types';
 import { useEffect, useState } from 'react';
+import type { TagType } from '../../../types/tag';
 import { initPerf } from '../../browser/initPerf';
 import { submitComponentEvent } from '../../browser/ophan/ophan';
 import {
