+++ dotcom-rendering/src/web/lib/ArticleRenderer.tsx
@@ -2,6 +2,7 @@ import { css } from '@emotion/react';
 import type { ArticleFormat } from '@guardian/libs';
 import { ArticleDesign } from '@guardian/libs';
 import type { Switches } from '../../types/config';
+import type { TagType } from '../../types/tag';
 import {
 	adCollapseStyles,
 	labelStyles as adLabelStyles,
