+++ dotcom-rendering/src/web/lib/useSignInGateWillShow.ts
@@ -1,4 +1,5 @@
 import { useEffect, useState } from 'react';
+import type { TagType } from '../../types/tag';
 import type {
 	CurrentSignInGateABTest,
 	SignInGateComponent,
