+++ dotcom-rendering/src/web/lib/withSignInGateSlot.tsx
@@ -3,6 +3,7 @@
 
 import React from 'react';
 import type { Switches } from '../../types/config';
+import type { TagType } from '../../types/tag';
 import { Island } from '../components/Island';
 import { SignInGateSelector } from '../components/SignInGateSelector.importable';
